improve tests

+++ src/graph.hpp
@@ -7,8 +7,8 @@
 class Graph {
   public:
     Graph(std::vector<int> vertices, std::vector<int> edges);
-    Graph(const std::string csvFilePath, const bool reverseColumns, const int numberOfVertices,
-          const int numberOFEdges);
+    Graph(const std::string csvFilePath, const bool reverseColumns, const size_t numberOfVertices,
+          const size_t numberOFEdges);
     std::vector<int> bfsCPU(int startVertex);
     std::vector<int> bfsGPU(int startVertex);
     ~Graph();
